um. well it should work, but it's...like...not
