+++ app/src/main/res/layout/activity_main.xml
@@ -42,25 +42,31 @@
 		    app:tabMode="scrollable" />
     </android.support.design.widget.AppBarLayout>
 
-	<android.support.v4.view.ViewPager
-		android:id="@+id/viewpager"
+	<android.support.v4.widget.SwipeRefreshLayout
 		android:layout_width="match_parent"
 		android:layout_height="match_parent"
-		app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+		android:id="@+id/swipeRefreshLayout">
 
-    <ProgressBar
-                android:id="@+id/loadingBar"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_centerHorizontal="true"
-				android:layout_centerVertical="true"/>
+		<android.support.v4.view.ViewPager
+			android:id="@+id/viewpager"
+			android:layout_width="match_parent"
+			android:layout_height="match_parent"
+			app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+	</android.support.v4.widget.SwipeRefreshLayout>
 
-    <TextView
-        android:id="@+id/status"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@id/loadingBar"
-        android:text="@string/loading"
-        android:textSize="16sp"/>
+	<ProgressBar
+		android:id="@+id/loadingBar"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_centerHorizontal="true"
+		android:layout_centerVertical="true"/>
+
+	<TextView
+		android:id="@+id/status"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_centerHorizontal="true"
+		android:layout_below="@id/loadingBar"
+		android:text="@string/loading"
+		android:textSize="16sp"/>
 </android.support.design.widget.CoordinatorLayout>
